Use long for Owner ID to be consistent

+++ backend/src/main/java/hu/bsstudio/raktr/dal/repository/OwnerRepository.java
@@ -3,7 +3,7 @@ package hu.bsstudio.raktr.dal.repository;
 import hu.bsstudio.raktr.dal.entity.Owner;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OwnerRepository extends JpaRepository<Owner, Integer> {
+public interface OwnerRepository extends JpaRepository<Owner, Long> {
 
     boolean existsByName(String name);
